requirements.txt
six
tqdm
torchtext>=0.2.1
future
networkx